+++ app/assets/javascripts/components/Calendar/DayCalendar/Plan.js
