+++ app/assets/javascripts/components/Calendar/MonthCalendar.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+class MonthCalendar extends React.Component {
+  render() {
+    return (
+      <div>Month Calendar</div>
+    );
+  }
+}
+
+export default MonthCalendar;
