+++ app/assets/javascripts/components/Calendar/WeekCalendar.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+class WeekCalendar extends React.Component {
+  render() {
+    return (
+      <div>Week Calendar</div>
+    );
+  }
+}
+
+export default WeekCalendar;
