+++ app/assets/javascripts/components/Calendar/YearCalendar.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+class YearCalendar extends React.Component {
+  render() {
+    return (
+      <div>Year Calendar</div>
+    );
+  }
+}
+
+export default YearCalendar;
